REFACTOR: 6주차 실습 context-api #19
