+++ week02_todoList/src/main.jsx
@@ -2,5 +2,10 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
 import "./index.css";
+import { TodoContextProvider } from "./context/TodoContext.jsx";
 
-createRoot(document.getElementById("root")).render(<App />);
+createRoot(document.getElementById("root")).render(
+  <TodoContextProvider>
+    <App />
+  </TodoContextProvider>
+);
